backend/flask-server/requirements.txt
aniso8601==7.0.0
anyio==3.4.0
ariadne==0.14.0
attrs==21.2.0
bcrypt==3.2.0
bidict==0.21.4
cffi==1.15.0
click==8.0.3
configparser==5.2.0
cryptography==36.0.1
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-GraphQL==2.0.1
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.2.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.2
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
passlib==1.7.4
promise==2.3
protobuf==3.19.1
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyrsistent==0.18.0
python-dotenv==0.19.2
python-engineio==4.3.0
python-socketio==5.5.0
pytz==2021.3
PyYAML==6.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.17.1
typing_extensions==4.0.1
uuid==1.30
Werkzeug==2.0.2
zipp==3.6.0
